+++ src/components/sections/ProductDetails/ProductDetails.tsx
@@ -194,6 +194,11 @@ function ProductDetails({ product: staleProduct }: Props) {
             secondOffer={mockedSubscriptionOffers.secondOffer}
           />
         </section>
+
+        <section className="product-questions__container">
+          <LatestQuestions productQuestions={productQuestions} />
+          <QuestionForm />
+        </section>
       </section>
     </Section>
   )