+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - **GM-397** - Filter mobile
+- **GM-398** - [PDP] Show latest questions list
+- **GM-108** - [PDP] Show common questions form
 
 ### Added